moved test file into test folder and transform api into lib folder working and added new scripts into the package JSON
