+++ .eslintrc
@@ -1,3 +1,4 @@
+ {
  "globals": {
    "describe": false,
    "it": false,
@@ -22,7 +23,6 @@
    "sourceType": "module",
    "ecmaFeatures": {"jsx": true}
  },
-
  "rules": {
    "semi": 1,
    "strict": 0,
