+++ lib/transform.js
